+++ ubo_app/rpc/ubo_bindings/client.py
@@ -3,22 +3,50 @@
 from __future__ import annotations
 
 import asyncio
-from typing import TYPE_CHECKING, overload
+from typing import TYPE_CHECKING, cast, overload
 
+from betterproto.lib.std.google import protobuf as betterproto_protobuf
 from grpclib.client import Channel
 
+import ubo_bindings.ubo.v1
 from ubo_bindings.secrets.v1 import QuerySecretRequest, SecretsServiceStub
 from ubo_bindings.store.v1 import (
     DispatchActionRequest,
     DispatchEventRequest,
     StoreServiceStub,
     SubscribeEventRequest,
+    SubscribeStoreRequest,
 )
 
 if TYPE_CHECKING:
     from collections.abc import Callable
 
+    from betterproto import Message
+
     from ubo_bindings.ubo.v1 import Action, Event
+
+
+def _unpack_from_any(message: betterproto_protobuf.Any) -> Message:
+    if not message.type_url.startswith('type.googleapis.com/'):
+        msg = f'Unsupported type URL: {message.type_url}'
+        raise ValueError(msg)
+
+    type_name = message.type_url[len('type.googleapis.com/') :]
+
+    if type_name.startswith('google.protobuf.'):
+        type_name = type_name[len('type.googleapis.com/google.protobuf.') :]
+        cls = cast('type[Message]', getattr(betterproto_protobuf, type_name, None))
+        if cls is not None:
+            return cls.FromString(message.value)
+
+    if type_name.startswith('ubo_bindings.ubo.v1.'):
+        type_name = type_name[len('ubo_bindings.ubo.v1.') :]
+        cls = cast('type[Message]', getattr(ubo_bindings.ubo.v1, type_name, None))
+        if cls is not None:
+            return cls.FromString(message.value)
+
+    msg = f'Unknown type URL: {message.type_url}'
+    raise ValueError(msg)
 
 
 class UboRPCClient:
@@ -71,6 +99,23 @@ class UboRPCClient:
                 callback(response.event)
 
         self.event_loop.create_task(iterator())
+
+    def autorun(
+        self,
+        selectors: list[str],
+    ) -> Callable[[Callable[[list], None]], None]:
+        """Autorun a function based on store changes."""
+
+        def wrapper(callback: Callable[[list], None]) -> None:
+            async def iterator() -> None:
+                async for response in self.store_service.subscribe_store(
+                    SubscribeStoreRequest(selectors=selectors),
+                ):
+                    callback([_unpack_from_any(item) for item in response.results])
+
+            self.event_loop.create_task(iterator())
+
+        return wrapper
 
     async def query_secret(
         self,